+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+
+## [0.0.3-dev] - Add new widget 
+
+* Add new widget ***ImageButton***
+* Add new widget ***Content***
+* Add new widget ***Calendar***
+* Add new style ***FlatStyle***
+
+## [0.0.2] - Add new method
+
+* add RWOrientationBuilder
+
 ## [0.0.1] - First publish
 
 * Just contains package create template.